Refactor function signatures, add move tests
Path function signatures now accept anything that can be converted to a
platform native string used for paths (i.e. wchar_t in Windows, char in
anything else). If the parameter needs to be converted to the platform
native string, a temporary buffer is allocated, written to, used by the
function and then deallocated.

Add move tests for moving directories